+++ lib-multisrc/paprika/src/eu/kanade/tachiyomi/multisrc/paprika/Filters.kt
@@ -1,0 +1,123 @@
+package eu.kanade.tachiyomi.multisrc.paprika
+
+import eu.kanade.tachiyomi.source.model.Filter
+
+open class UriPartFilter(
+    displayName: String,
+    private val vals: Array<Pair<String, String>>,
+) : Filter.Select<String>(
+    displayName,
+    vals.map { it.first }.toTypedArray(),
+) {
+    fun toUriPart() = vals[state].second
+}
+
+class OrderFilter : UriPartFilter("Order by", ORDER_LIST)
+class GenreFilter : UriPartFilter("Genre", GENRE_LIST)
+
+private val ORDER_LIST = arrayOf(
+    Pair("Views", "2"),
+    Pair("Latest", "3"),
+    Pair("A-Z", "1"),
+)
+
+private val GENRE_LIST = arrayOf(
+    Pair("4 koma", "4-koma"),
+    Pair("Action", "action"),
+    Pair("Adaptation", "adaptation"),
+    Pair("Adult", "adult"),
+    Pair("Adventure", "adventure"),
+    Pair("Aliens", "aliens"),
+    Pair("Animals", "animals"),
+    Pair("Anthology", "anthology"),
+    Pair("Award winning", "award-winning"),
+    Pair("Comedy", "comedy"),
+    Pair("Cooking", "cooking"),
+    Pair("Crime", "crime"),
+    Pair("Crossdressing", "crossdressing"),
+    Pair("Delinquents", "delinquents"),
+    Pair("Demons", "demons"),
+    Pair("Doujinshi", "doujinshi"),
+    Pair("Drama", "drama"),
+    Pair("Ecchi", "ecchi"),
+    Pair("Fan colored", "fan-colored"),
+    Pair("Fantasy", "fantasy"),
+    Pair("Food", "food"),
+    Pair("Full color", "full-color"),
+    Pair("Game", "game"),
+    Pair("Gender bender", "gender-bender"),
+    Pair("Genderswap", "genderswap"),
+    Pair("Ghosts", "ghosts"),
+    Pair("Gore", "gore"),
+    Pair("Gossip", "gossip"),
+    Pair("Gyaru", "gyaru"),
+    Pair("Harem", "harem"),
+    Pair("Historical", "historical"),
+    Pair("Horror", "horror"),
+    Pair("Isekai", "isekai"),
+    Pair("Josei", "josei"),
+    Pair("Kids", "kids"),
+    Pair("Loli", "loli"),
+    Pair("Lolicon", "lolicon"),
+    Pair("Long strip", "long-strip"),
+    Pair("Mafia", "mafia"),
+    Pair("Magic", "magic"),
+    Pair("Magical girls", "magical-girls"),
+    Pair("Manhwa", "manhwa"),
+    Pair("Martial arts", "martial-arts"),
+    Pair("Mature", "mature"),
+    Pair("Mecha", "mecha"),
+    Pair("Medical", "medical"),
+    Pair("Military", "military"),
+    Pair("Monster girls", "monster-girls"),
+    Pair("Monsters", "monsters"),
+    Pair("Music", "music"),
+    Pair("Mystery", "mystery"),
+    Pair("Ninja", "ninja"),
+    Pair("Office workers", "office-workers"),
+    Pair("Official colored", "official-colored"),
+    Pair("One shot", "one-shot"),
+    Pair("Parody", "parody"),
+    Pair("Philosophical", "philosophical"),
+    Pair("Police", "police"),
+    Pair("Post apocalyptic", "post-apocalyptic"),
+    Pair("Psychological", "psychological"),
+    Pair("Reincarnation", "reincarnation"),
+    Pair("Reverse harem", "reverse-harem"),
+    Pair("Romance", "romance"),
+    Pair("Samurai", "samurai"),
+    Pair("School life", "school-life"),
+    Pair("Sci fi", "sci-fi"),
+    Pair("Seinen", "seinen"),
+    Pair("Shota", "shota"),
+    Pair("Shotacon", "shotacon"),
+    Pair("Shoujo", "shoujo"),
+    Pair("Shoujo ai", "shoujo-ai"),
+    Pair("Shounen", "shounen"),
+    Pair("Shounen ai", "shounen-ai"),
+    Pair("Slice of life", "slice-of-life"),
+    Pair("Smut", "smut"),
+    Pair("Space", "space"),
+    Pair("Sports", "sports"),
+    Pair("Super power", "super-power"),
+    Pair("Superhero", "superhero"),
+    Pair("Supernatural", "supernatural"),
+    Pair("Survival", "survival"),
+    Pair("Suspense", "suspense"),
+    Pair("Thriller", "thriller"),
+    Pair("Time travel", "time-travel"),
+    Pair("Toomics", "toomics"),
+    Pair("Traditional games", "traditional-games"),
+    Pair("Tragedy", "tragedy"),
+    Pair("User created", "user-created"),
+    Pair("Vampire", "vampire"),
+    Pair("Vampires", "vampires"),
+    Pair("Video games", "video-games"),
+    Pair("Virtual reality", "virtual-reality"),
+    Pair("Web comic", "web-comic"),
+    Pair("Webtoon", "webtoon"),
+    Pair("Wuxia", "wuxia"),
+    Pair("Yaoi", "yaoi"),
+    Pair("Yuri", "yuri"),
+    Pair("Zombies", "zombies"),
+)
